Write authors as array in PDF 2.0
Currently, authors are always folded into a comma-separated single entry in XMP, even though it supports an array. This is a compatibility behavior. The background behind this is that the PDF Document Info dictionary cannot express multiple authors but XMP can. When a document has both present and they do not match, Acrobat will first render the DocInfo author and then the 2nd , 3rd, ... author from the XMP.

In PDF 2.0, the whole document info dictionary has been deprecated in favor of XMP. That means that for these documents, we can disable this compatibility behavior and write an array.