Map `Metadata::subject` to `dc:description`
This PR changes how `Metadata::subject` is serialized to align with Typst 0.13 behavior and interoperability guidance.

In PDF 1.1, the Subject field in the Document Information Dictionary is defined as "The subject of the document." ISO 19005-1 (PDF/A-1) specified that this should match the contents of `dc:subject`. However, ISO published Technical Corrigendum 1 in 2007 to change this to `dc:description["x-default"]`. The contents of this erratum are incorporated in the [PDF Association's Technical Note 0003 – Metadata in PDF/A-1](https://pdfa.org/resource/technical-note-tn0003-metadata-in-pdfa-1/). In ISO 19005-2 and 19005-3, the crosswalk is also defined as `dc:description["x-default"]`.

This is supported by the [Dublin Core RFC](https://www.rfc-editor.org/rfc/rfc5013.html). In Section 6, it defines the elements like this:

- `dc.subject`:
  > Typically, the subject will be represented using keywords, key phrases, or classification codes. Recommended best practice is to use a controlled vocabulary.  To describe the spatial or temporal topic of the resource, use the Coverage element.
- `dc:description`:
  > Description may include but is not limited to: an abstract, a table of contents, a graphical representation, or a free-text account of the resource.

Accordingly, XMP represents `dc:subject` as an unordered array of text and `dc:description` as a language alternative (like `dc:title`).

Due to this semantic meaning, [Acrobat and Foxit PDF merge `dc:subject` into the keywords](https://github.com/typst/typst/issues/7017).

I am not aware of krilla dependents that use `Metadata::set_subject` with these semantics.

I'd also like to propose to change the name of `Metadata::subject` to `Metadata::description`. In lieu of this change, I have added more guidance to the doc comment.