Add Separation color space
This commit adds support for the special `Separation` color space. It exposes the colorant name in the `Separation` dictionary, including typed writing for the `All` and `None` colorants. For fallback, the color space accepts a color that is then linearly interpolated given the tint, so both the `alternateSpace` and `tintTransform` array entries are managed by Krilla.

On standards compliance: PDF/A-2 and later require the same colorants to always come with the same `alternateSpace` and `tintTransform` entries. I am tracking this in the new `ValidationStore` struct. PDF/UA has no special requirements for separation color spaces.

I had to remove the `Copy` impl from Color because Separations can contain strings. Let me know if e.g. `separation::Color` should contain a reference instead or `SeparationColorant` a Cow instead of a plain string.

I have added snapshot tests and manually inspected the result in a PDF reader.